+++ reactjs/src/components/molecules/challenges/ChallengeNavbar.tsx
@@ -51,7 +51,7 @@ const ChallengeNavbar = (props: any) => {
         onClose={handleClose}
       >
         <NavItem to={`/challenges/${props.id}/timeline`} text="タイムライン" />
-        <NavItem to={`/challenges/${props.id}/discussion`} text="掲示板" />
+        <NavItem to={`/challenges/${props.id}/discussion`} text="知見まとめ" />
         <NavItem
           to={`/challenges/${props.id}/leaderboard`}
           text="リーダーボード"
